Improve gamedir & config dir detection
