+++ src/main/kotlin/online/ruin_of_future/informative_mc_core/config/ModConfig.kt
@@ -16,8 +16,8 @@
 package online.ruin_of_future.informative_mc_core.config
 
 import kotlinx.serialization.Serializable
-import online.ruin_of_future.informative_mc_core.util.generateRandomString
 import online.ruin_of_future.informative_mc_core.modConfigDirPath
+import online.ruin_of_future.informative_mc_core.util.generateRandomString
 import java.io.File
 
 @Serializable
