+++ src/main/kotlin/online/ruin_of_future/informative_mc_core/web_api/handler/UserRegister.kt
@@ -18,8 +18,8 @@ package online.ruin_of_future.informative_mc_core.web_api.handler
 import kotlinx.serialization.Serializable
 import online.ruin_of_future.informative_mc_core.data.ImcUser
 import online.ruin_of_future.informative_mc_core.data.ModData
-import online.ruin_of_future.informative_mc_core.util.UUIDSerializer
 import online.ruin_of_future.informative_mc_core.token_system.TokenManager
+import online.ruin_of_future.informative_mc_core.util.UUIDSerializer
 import online.ruin_of_future.informative_mc_core.web_api.ApiID
 import java.io.OutputStream
 import java.util.*
